+++ src/products-inventory/category/dto/category-response.dto.ts
@@ -1,5 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
+import { SuccessResponse } from 'src/common/dtos/success-response.dto';
 import { MerchantResponseDto } from 'src/merchants/dtos/merchant-response.dto';
+import { Category } from '../entities/category.entity';
 
 export class CategoryResponseDto {
   @ApiProperty({ example: 1, description: 'Category ID' })
@@ -38,4 +40,27 @@ export class CategoryResponseDto {
     nullable: true,
   })
   parents?: { id: number; parentName: string }[];
+}
+export class CategoryLittleResponseDto {
+  @ApiProperty({ example: 1, description: 'Category ID' })
+  id: number;
+
+  @ApiProperty({ example: 'Beverages', description: 'Category name' })
+  name: string;
+
+  @ApiProperty({
+    type: () => CategoryLittleResponseDto,
+    nullable: true,
+    description: 'Parent category details',
+  })
+  parent?: CategoryLittleResponseDto | null;
+}
+
+export class OneCategoryResponse extends SuccessResponse {
+  @ApiProperty()
+  data: Category;
+}
+export class AllCategoryResponse extends SuccessResponse {
+  @ApiProperty()
+  data: Category[];
 }
